requirements.txt
anyjson==0.3.3
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
boto==2.40.0
CD==1.1.0
coverage==4.0.2
decorator==4.0.11
Django==1.11
django-appconf==1.0.1
django-compressor==2.1
django-easy-timezones==0.7.0
django-hitcount==1.2.2
django-image-cropping==1.0.4
django-redis==4.4.4
django-redis-cache==1.7.0
django-storages==1.4.1
easy-thumbnails==2.4.1
enum34==1.1.6
filemagic==1.6
funcsigs==1.0.2
geopy==1.11.0
google-api-python-client==1.5.3
httplib2==0.9.2
ipaddress==1.0.14
ipython==5.3.0
ipython-genutils==0.2.0
mock==2.0.0
MySQL-python==1.2.5
numpy==1.8.0
oauth2client==3.0.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pygeoip==0.3.2
Pygments==2.2.0
PyJWT==1.4.2
PyMySQL==0.6.7
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
rcssmin==1.0.6
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
rjsmin==1.0.12
rsa==3.4.2
scandir==1.5
shortuuid==0.4.2
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
traitlets==4.3.2
uritemplate==0.6
wcwidth==0.1.7
